update: username field to match authenticated driver from keycloak

+++ driver-service/src/main/resources/db/changelog/v-1.0/20-create-driver-rating-table.sql
@@ -3,7 +3,7 @@ CREATE TABLE driver_rating
     id             BIGSERIAL PRIMARY KEY,
     rating         INTEGER NOT NULL,
     feedback       TEXT,
-    driver_info_id INTEGER,
+    driver_info_id BIGINT,
     CONSTRAINT fk_driver_info
         FOREIGN KEY (driver_info_id)
             REFERENCES driver_info (id)
